+++ src/ep2/Map.java
@@ -37,7 +37,7 @@ public class Map extends MapReduceBase implements
 			dps.adicionaDano(dmg.getSource(), dmg.getAmount(), dmg.getTime());
 		}
 		word.set("");
-		output.collect(word, new DoubleWritable((long)dps.geraEstatistica("")));
+		output.collect(word, new DoubleWritable(dps.geraEstatistica("")));
 	}
 
 }
