+++ src/ep2/Reduce.java
@@ -6,7 +6,7 @@ package ep2;
 import java.io.IOException;
 import java.util.Iterator;
 
-import org.apache.hadoop.io.IntWritable;
+import org.apache.hadoop.io.DoubleWritable;
 import org.apache.hadoop.io.Text;
 import org.apache.hadoop.mapred.MapReduceBase;
 import org.apache.hadoop.mapred.OutputCollector;
@@ -18,17 +18,17 @@ import org.apache.hadoop.mapred.Reporter;
  *
  */
 public class Reduce extends MapReduceBase implements
-		Reducer<Text, IntWritable, Text, IntWritable> {
+		Reducer<Text, DoubleWritable, Text, DoubleWritable> {
 
 	@Override
-	public void reduce(Text key, Iterator<IntWritable> values,
-			OutputCollector<Text, IntWritable> output, Reporter reporter)
+	public void reduce(Text key, Iterator<DoubleWritable> values,
+			OutputCollector<Text, DoubleWritable> output, Reporter reporter)
 			throws IOException {
 		int sum = 0;
 		while (values.hasNext()) {
 			sum += values.next().get();
 		}
-		output.collect(key, new IntWritable(sum));
+		output.collect(key, new DoubleWritable(sum));
 	}
 
 }
